Various updates including a working model.
Added a test.php file which generates random orders and plays them out.

Currently running on my Macbook pro at around 260 orders processed per
second. My target is around 1500 per second in a production environment.

I enabled threading and seems to be handling it fairly well.

There is a possible memory leak, after around 50,000 orders I see
memory useage up to around 5-6 megs, but starts at about 1 meg. This
might not be a leak because there might be many large orders at at
either end of the randomly generated pricing.

Next on the list is to get it working with a nosql database. The speed
will likely slow down a bit when this is enabled :-(